[APIE-1098] Relocate the Flink package helpers to fit code generation
Two moves, both aimed at making `internal/flink` easier to generate into.

CMF helpers out of command.go. `addCmfFlagSet`, `addPageSizeFlag`,
`getPageSize`, `createContext` and `handleFlinkWebUiForwardRequest` are
on-prem-only — every caller is a CP command, including the `addCmfFlagSet`
call in command_shell.go, which sits in that command's on-prem arm. They now
live in cmf_utils_onprem.go, together with `copyDataType` from utils.go
(also CP-only, one caller), which lets utils.go go away. command.go drops
from 176 lines to 96 and sheds five imports; it is the file merge mode
writes into, so unrelated plumbing does not belong there.

`addComputePoolFlag` up to pkg/cmd/flags.go as `AddComputePoolFlag`, with
`AutocompleteComputePools` alongside it, modeled on AddClusterFlag /
AutocompleteClusters. The generator emits a flag registration call for a
`compute_pool` parameter, and that call has to work from any receiver; a
method on the flink package's `command` type only works for hand-written
flink commands. No suite emits it yet, so this is a latent break rather than
a live one — `flink statement` is the command that would have hit it.

Positional <id> completion stays a self-contained local method, and is
reshaped to match what the generator emits: `validComputePoolArgs` delegates
to `validComputePoolArgsMultiple`, which runs the prerun and calls a
no-argument `autocompleteComputePools()`. It duplicates the pkg helper's
loop on purpose — flag completion and positional completion are separate
paths, and matching the generated shape keeps the parity diff small.
