+++ internal/flink/command_statement_resume.go
@@ -45,7 +45,7 @@ func (c *command) newStatementResumeCommand() *cobra.Command {
 	}
 
 	c.addPrincipalFlag(cmd)
-	c.addComputePoolFlag(cmd)
+	pcmd.AddComputePoolFlag(cmd, c.AuthenticatedCLICommand)
 	pcmd.AddCloudFlag(cmd)
 	pcmd.AddRegionFlagFlink(cmd, c.AuthenticatedCLICommand)
 	pcmd.AddEnvironmentFlag(cmd, c.AuthenticatedCLICommand)
